HDMI works now.  I just needed to attach the OBUF directly to the pin. I already fixed this bug, but introduced a new one at the same time (recoded and somehow dropped the clock from the ODDR), so I didn't realize that was the correcct fix.  I guess it was.
While I was at it, I renamed the pins to HDMI.  Let them come and sue
me.  I added a 4th HDMI_TX pin as an alias to the clock so I can use
an array instead of recoding it (as mentioned above).

The demo could use some work to look prettier, but I don't care.  It's
just a test to see if the board actually works.